requirements.txt
# ════════════════════════════════════════════════════════════════════════════
# requirements.txt — Météo IA France V1.0
# ════════════════════════════════════════════════════════════════════════════
#
# 🎯 RÔLE
#   Packages Python à installer via PIP, APRÈS avoir créé l'env conda
#   avec environment.yml. Ces packages ne sont pas disponibles sur conda-forge
#   ou nécessitent une installation pip spécifique (GraphCast depuis Git).
#
# 📋 INSTALLATION (PRÉ-REQUIS : env conda déjà créé)
#
#   1. S'assurer que l'env conda est activé :
#      conda activate meteo_ia
#
#   2. Installer les packages pip :
#      pip install -r requirements.txt
#
#   3. Installer GraphCast séparément (n'est PAS sur PyPI) :
#      pip install git+https://github.com/deepmind/graphcast.git@08cf73625c9d12bd9aaa038868bcb2fe488f2a22
#
# 🐍 Python : 3.11 (déjà installé via conda)
# 📅 Dernière mise à jour : 29/04/2026
# 📦 ~70 packages pip
#
# ════════════════════════════════════════════════════════════════════════════
# ⚠️ NOTES IMPORTANTES
# ════════════════════════════════════════════════════════════════════════════
#
# - Versions copiées de l'env conda meteo_ia de référence (29/04/2026)
# - Versions PINNÉES strictement (==) pour reproductibilité 100%
# - Pour upgrade : modifier ce fichier, retester complètement avant prod
# - Si problème "version not found" : vérifier que conda env est bien activé
#
# ════════════════════════════════════════════════════════════════════════════


# ─── ☁️ ACQUISITION DONNÉES MÉTÉO (APIs publiques) ──────────────────────────
# Climate Data Store (Copernicus) pour ERA5
cdsapi==0.7.7
ecmwf-datastores-client==0.5.1   # Transitif cdsapi
multiurl==0.3.7                  # Transitif cdsapi


# ─── ☁️ GOOGLE CLOUD STORAGE (modèle GraphCast + GFS NOMADS) ────────────────
gcsfs==2026.3.0
google-api-core==2.30.3
google-auth==2.49.2
google-auth-oauthlib==1.3.1
google-cloud-core==2.5.1
google-cloud-storage==3.10.1
google-cloud-storage-control==1.11.0
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpc-google-iam-v1==0.14.4
grpcio==1.80.0
grpcio-status==1.80.0
proto-plus==1.27.2
protobuf==6.33.6


# ─── 🤖 GRAPHCAST DEPENDENCIES (couches sur JAX) ────────────────────────────
# IMPORTANT : jax + jaxlib 0.9.2 viennent de CONDA, pas de pip
# Ces packages-ci sont les deps PYTHON PURES de GraphCast
chex==0.1.91
dm-haiku==0.0.16
dm-tree==0.1.10
jmp==0.0.4                       # Mixed precision JAX
jraph==0.0.6.dev0                # Graph neural networks JAX (DeepMind)
trimesh==4.11.5                  # Mesh 3D pour graph GraphCast
tensorstore==0.1.82              # Stockage tensors GraphCast (checkpoints)
xarray-tensorstore==0.3.0        # Pont xarray ↔ tensorstore
zarr==3.1.6                      # Stockage chunked arrays
numcodecs==0.16.5                # Codecs zarr

# GraphCast lui-même : à installer SÉPARÉMENT (cf. instructions header)
# pip install git+https://github.com/deepmind/graphcast.git@08cf73625c9d12bd9aaa038868bcb2fe488f2a22


# ─── 🗄️ BASE DE DONNÉES POSTGRESQL ─────────────────────────────────────────
# Driver Postgres avec wheels précompilés (V1.0 simplification)
# En V1.1 prod stricte : remplacer par psycopg2 (compilé, libpq-dev requis)
psycopg2-binary==2.9.12


# ─── 🔧 UTILITAIRES PYTHON ──────────────────────────────────────────────────
python-dotenv==1.2.2             # Lecture des fichiers .env
tqdm==4.67.3                     # Barre de progression
cloudpickle==3.1.2               # Sérialisation Python avancée (deps JAX)
pytz==2026.1.post1               # Fuseaux horaires
pyyaml==6.0.3                    # YAML (configs)


# ─── 🌐 HTTP & ASYNC ────────────────────────────────────────────────────────
requests-oauthlib==2.0.0
oauthlib==3.3.1

# HTTP async (utilisé par gcsfs)
aiohttp==3.13.5
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
multidict==6.7.1
yarl==1.23.0
frozenlist==1.8.0
propcache==0.4.1


# ─── 📊 OUTILS DATA SCIENCE COMPLÉMENTAIRES ─────────────────────────────────
# Calcul énergie solaire (pour debug TOA radiation et futur PV V2.0)
pvlib==0.15.1

# Machine learning utilitaires
scikit-learn==1.8.0
joblib==1.5.3
threadpoolctl==3.6.0

# HDF5 Python (deps GraphCast pour les checkpoints)
h5py==3.16.0

# Calcul parallèle (xarray dépend de dask pour le lazy loading)
dask==2026.3.0
fsspec==2026.3.0
toolz==1.1.0
locket==1.0.0
partd==1.4.2

# Indexation spatiale (utile pour requêtes géo PostGIS)
rtree==1.4.1

# Tableaux formatés pour debug CLI
tabulate==0.10.0


# ─── 📚 DEPENDANCES TRANSITIVES (pinned pour reproductibilité 100%) ─────────
# Ces packages sont des deps transitives. On les pinne explicitement pour
# garantir que l'env est exactement reproductible sur le VPS.
absl-py==2.4.0
cryptography==46.0.7
decorator==5.2.1
typing-extensions==4.15.0
wrapt==2.1.2
platformdirs==4.9.6
pyasn1==0.6.3
pyasn1-modules==0.4.2
flexcache==0.3
flexparser==0.4
pint==0.25.3
donfig==0.8.1.post1
colabtools==0.0.1
dinosaur-dycore==1.2.1
tree-math==0.2.1
